requirements.txt
streamlit
matplotlib
numpy
pyrebase4
